requirements.txt
annotated-types==0.7.0
anyio==4.4.0
azure-ai-formrecognizer==3.3.3
azure-common==1.1.28
azure-core==1.30.2
azure-cosmos==4.7.0
azure-functions==1.20.0
azure-storage-blob==12.21.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
distro==1.9.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
isodate==0.6.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.2.21
langchain-openai==0.1.17
langchain-text-splitters==0.2.2
langsmith==0.1.92
msrest==0.7.1
oauthlib==3.2.2
openai==1.35.15
orjson==3.10.6
packaging==24.1
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
sniffio==1.3.1
tenacity==8.5.0
tiktoken==0.7.0
tqdm==4.66.4
typing_extensions==4.12.2
urllib3==2.2.2
